backend/requirements.txt
Flask==2.2.3
Flask-Cors==3.0.10
psycopg2-binary==2.9.5
python-dotenv==1.0.0
gunicorn==20.1.0
Werkzeug==2.2.3
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.3
